requirements.txt
pandas
numpy
openai
plotly
seaborn
matplotlib
pyyaml
fastapi
uvicorn
concurrent.futures
